operate-enhanced/pyproject.toml
[tool.poetry]
name = "operate-enhanced"
version = "2.0.0"
description = "Enhanced Self-Operating Computer Framework with GitHub integration"
authors = ["Enhanced Team"]
readme = "README.md"
python = "^3.9"

[tool.poetry.dependencies]
python = "^3.9"
pyautogui = "^0.9.54"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
openai = "^1.0.0"
anthropic = "^0.8.0"
google-generativeai = "^0.3.0"
easyocr = "^1.7.0"
pyyaml = "^6.0"
pydantic = "^2.5.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
websockets = "^12.0"
pygithub = "^2.1.0"
gitpython = "^3.1.40"
redis = "^5.0.0"
aiofiles = "^23.2.0"
asyncio = "^3.4.3"
httpx = "^0.25.0"
rich = "^13.7.0"
click = "^8.1.7"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
operate = "operate.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C90", "D"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true